// src/types/api.ts
export interface Attachment {
  name: string
  filename: string
  mimeType: string
  size: number
  addedAt: string
}